+++ src/constant/options.js
@@ -7,7 +7,10 @@ const OPTIONS = {
     finalWinnerDelimiter: ', '
   },
   CAR: {
+    minNameLength: 1,
     maxNameLength: 5,
     leastMoveCondition: 4
   }
 };
+
+export default OPTIONS;
